Ghibli-Food-ML/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
Pillow==10.1.0
nltk==3.8.1
textblob==0.17.1
joblib==1.3.2
seaborn==0.13.0
matplotlib==3.8.2
plotly==5.17.0
jupyter==1.0.0
notebook==7.0.6
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4